2.1: Start to decode the data inside the Chain State and formatting code
This commit include the initial logic read the first two value of the level db.
For the moment SpyCBlock is able to decode the transaction id and the v out
In particular the Vout in under osservation because it use the Bitcoin Core APi to read conver the VarInt into integer, but to do this, I need to make a simple trick with stream. For this reason this feature is unsafe at the moment.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>